+++ components/data/data-processes/src/main/java/org/eclipse/dirigible/components/data/processes/schema/export/tasks/BaseExportTask.java
@@ -1,7 +1,16 @@
 package org.eclipse.dirigible.components.data.processes.schema.export.tasks;
 
+import org.eclipse.dirigible.components.base.helpers.JsonHelper;
 import org.eclipse.dirigible.components.engine.bpm.flowable.delegate.BPMTask;
 import org.eclipse.dirigible.components.engine.bpm.flowable.delegate.TaskExecution;
+import org.eclipse.dirigible.components.engine.cms.*;
+import org.springframework.http.MediaType;
+
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
+import java.util.Map;
 
 abstract class BaseExportTask extends BPMTask {
 
@@ -13,4 +22,52 @@ abstract class BaseExportTask extends BPMTask {
 
     protected abstract void execute(ExportProcessContext context);
 
+    protected void saveObjectAsJsonDocument(Object object, String fileName, String folderPath) {
+        String exportTopologyJson = JsonHelper.toJson(object);
+        saveDocument(exportTopologyJson, fileName, MediaType.APPLICATION_JSON_VALUE, folderPath);
+    }
+
+    private void saveDocument(String content, String fileName, String mediaType, String folderPath) {
+        CmisFolder folder = getFolder(folderPath);
+        saveDocument(content, fileName, mediaType, folder);
+    }
+
+    private void saveDocument(String content, String fileName, String mediaType, CmisFolder folder) {
+        CmisSession cmisSession = CmisSessionFactory.getSession();
+
+        Map<String, String> fileProps = Map.of(CmisConstants.OBJECT_TYPE_ID, CmisConstants.OBJECT_TYPE_DOCUMENT, //
+                CmisConstants.NAME, fileName);
+
+        // this implementation is fine for small files
+        byte[] bytes = content.getBytes(StandardCharsets.UTF_8);
+        try (InputStream inStream = new ByteArrayInputStream(bytes)) {
+
+            long length = bytes.length; // not needed?
+            CmisContentStream contentStream = cmisSession.getObjectFactory()
+                                                         .createContentStream(fileName, length, mediaType, inStream);
+
+            folder.createDocument(fileProps, contentStream);
+        } catch (IOException ex) {
+            throw new SchemaExportException("Failed to create document with name [" + fileName + "] in folder [" + folder + "]", ex);
+        }
+    }
+
+    protected CmisFolder getFolder(String folderPath) {
+        CmisSession cmisSession = CmisSessionFactory.getSession();
+
+        try {
+            CmisObject exportFolderAsObject = cmisSession.getObjectByPath(folderPath);
+            if (exportFolderAsObject instanceof CmisFolder exportFolder) {
+                return exportFolder;
+            }
+            throw new SchemaExportException("Returned cmis object " + exportFolderAsObject + " is not a folder");
+        } catch (IOException ex) {
+            throw new SchemaExportException("Failed to get folder from path " + folderPath, ex);
+        }
+    }
+
+    protected void saveObjectAsJsonDocument(Object object, String fileName, CmisFolder folder) {
+        String exportTopologyJson = JsonHelper.toJson(object);
+        saveDocument(exportTopologyJson, fileName, MediaType.APPLICATION_JSON_VALUE, folder);
+    }
 }
